Add STLs for engraved top cases
These are already a part of the included Fusion 360 file, but suppressed at the end of the feature tree. At the time, adding additional STLs that were identical to the existing "Top" STL in everything but a debossed logo seemed redundant. Coming back to this two years later to print a case and I realize it would be nice to have these available so that users don't have to install F360 and export them. Obvious, in hindsight...