Bill of Materials (BOM) (#5)
Adds Bill of Materials (BOM) information to the KiCad schematic, which can be exported to CSV using the "Generate Bill of Materials" option in the toolbar of the schematic editor.

These parts are, to the best of my knowledge, the parts I used when assembling these boards two and a half years ago. You do not need to use these exact parts, substitutes may be available that are currently stocked and/or less expensive. The 1% resistors, in particular, do not need to be tight tolerance - that's just what I had on hand.