Shifter: Move power to pin 9 (swap 7/9)
This is an error on the v1 PCB, and the fundamental reason why this PCB doesn't work with the G25 shifter.

On the Driving Force shifter, pins 7 and 9 are connected together internally in the DE-9 connector. Power applied to one goes to both pins, so one is used to power the shifter and the other is used with a pull-down to act as a detection pin. Since it didn't seem to matter I arbitrarily chose pin 7 as the power input for easier trace routing.

On the G25 shifter, pins *1* and 9 are connected together internally in the DE-9 connector. Pin 7 is instead attached to the clock input of the EEPROM / shift registers, and needs to be pulsed in order to get the button data from the shifter.

50/50 and I lost. Ah well.