Shifter Shield Version 2 (G25 + G27) (#6)
* Shifter: Changed all jumpers to NC

We need all of these pins in order to interface with the Logitech G25 shifter

* Shifter: Move power to pin 9 (swap 7/9)

This is an error on the v1 PCB, and the fundamental reason why this PCB doesn't work with the G25 shifter.

On the Driving Force shifter, pins 7 and 9 are connected together internally in the DE-9 connector. Power applied to one goes to both pins, so one is used to power the shifter and the other is used with a pull-down to act as a detection pin. Since it didn't seem to matter I arbitrarily chose pin 7 as the power input for easier trace routing.

On the G25 shifter, pins *1* and 9 are connected together internally in the DE-9 connector. Pin 7 is instead attached to the clock input of the EEPROM / shift registers, and needs to be pulsed in order to get the button data from the shifter.

50/50 and I lost. Ah well.

* Shifter: Schematic to support G25

Adds a pull-down resistor on the second 'detect' pin, and adds three series resistors to protect against shorts between the two shifter types (Driving Force / G25).

Also renames all DE9 connections with Shift_{G25}_{G29} labels.

* Shifter: Update bottom silkscreen files

* Shifter: Modify PCB for G25

* Add firmware build files to gitignore

* Shifter: Add firmware

Supports the Driving Force (G293 / G920 / G29), G27, and G25 shifters interchangeably.

* Shifter: Add G27 notes to schematic

Lucked out that the G25 hardware changes also support the G27.

* Modify README for shifter v2